requirements.txt
pyqt6>=6.6.0
pillow>=10.2.0
pillow-heif>=0.15.0
send2trash>=1.8.0
psutil>=5.9.0
